extract functions to helper
